indexer/requirements.txt
openai==1.63.0
httpx==0.27.2
faiss-cpu==1.8.0
numpy==1.26.4
python-dotenv==1.0.1
pydantic-settings==2.3.4
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==4.2.0
python-docx==1.1.0
markdown==3.6
tqdm